refactor(connectors): isolate Hermes shim lifecycle; persist config.yaml provenance
Sixth review batch for PR #1938:

- All Hermes discovery-contract logic (home resolution, discovery
  heuristic, shim content/marker, reconcile/remove lifecycle) moves to
  a dedicated connectors/hermes-shim.ts module, following the
  codex-materialize.ts precedent, so the generic connector registry
  only orchestrates (Codex P1). Full relocation into the remnic-hermes
  package is not possible — that package is Python, and the npm CLI
  must be able to materialize the shim without a Python runtime; the
  connectors subsystem is the established home for per-host installer
  logic (codex, weclone, hermes).
- Install now persists hermesConfigPath (the config.yaml written under
  the active HERMES_HOME) alongside pluginShimPath, and
  removeHermesConfig accepts shape-guarded extraConfigPaths, so a
  HERMES_HOME change between install and remove cannot leave the old
  home's remnic: block (with its daemon token) behind (Codex P2).

Cross-HERMES_HOME removal test now also asserts the persisted
config.yaml cleanup (92 pass).

Part of #1929
